Comprobación de las preferences desde Launcher y añadido item para cerrar sesión
