--- App/app/src/main/java/com/afundacion/gestorfinanzasdesarrollointerfaces/Screens/FragmentExample.java
@@ -1,39 +1,0 @@
-package com.afundacion.gestorfinanzasdesarrollointerfaces.Screens;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import com.afundacion.gestorfinanzasdesarrollointerfaces.R;
-
-public class FragmentExample extends Fragment {
-
-    private static final String TEXT = "text";
-
-    public static FragmentExample newInstance(String text) {
-        FragmentExample frag = new FragmentExample();
-
-        Bundle args = new Bundle();
-        args.putString(TEXT, text);
-        frag.setArguments(args);
-
-        return frag;
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable
-            Bundle savedInstanceState) {
-        View layout = inflater.inflate(R.layout.fragment_example, container, false);
-
-        if (getArguments() != null) {
-            ((TextView) layout.findViewById(R.id.text)).setText(getArguments().getString(TEXT));
-        }
-
-        return layout;
-    }
-}
