+++ App/app/src/main/res/drawable/ic_menu_exit.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="48dp"
+    android:height="48dp"
+    android:viewportWidth="48"
+    android:viewportHeight="48">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M9,42q-1.2,0 -2.1,-0.9Q6,40.2 6,39L6,9q0,-1.2 0.9,-2.1Q7.8,6 9,6h14.55v3L9,9v30h14.55v3ZM33.3,32.75 L31.15,30.6 36.25,25.5h-17.5v-3h17.4l-5.1,-5.1 2.15,-2.15 8.8,8.8Z"/>
+</vector>
