+++ App/app/src/main/res/menu/drawer.xml
@@ -25,6 +25,11 @@
             android:icon="@drawable/ic_menu_estadisticas"
             android:title="Estadisticas"
             />
+
+        <item
+            android:id="@+id/nav_logout"
+            android:icon="@drawable/ic_menu_exit"
+            android:title="Cerrar sesión"/>
     </group>
 
 </menu>
